FIX loss of precision error when casting from pointer directly to uint. Did not want to change the method signature so the loss of precision is forced by introducing a cast to ulong

+++ application/Core/FileManager/priv/WordIndex/Node.h
@@ -122,7 +122,7 @@ namespace FM
    template <typename T>
    inline uint qHash(const NodeResult<T>& r)
    {
-      return uint(r.value);
+      return uint((ulong)r.value);
    }
 }
